move emulator data getters to new cpp
