+++ app/emulator/CMakeLists.txt
@@ -17,6 +17,7 @@ add_executable(${PROJECT_NAME}
     main.cpp
     src/emulator.cpp
     src/engine_pindle_states.cpp
+    src/emulator_data.cpp
 )
